requirements.txt
Scrapy==2.11.2
selenium==4.27.1
psycopg2==2.9.9
python-dotenv==1.0.1
pandas==2.0.3
openpyxl==3.1.5
sqlalchemy==2.0.44
aiosmtplib==5.0.0